+++ src/Components/AboutMe/AboutMe.component.scss
@@ -6,10 +6,15 @@
     border-radius: 20px;
     box-shadow: 0 4px 8px rgba(0, 0, 0, 0.5);
     width: auto;
-    height: 60vh;
+    height: 70vh;
   }
   &_text {
     height: auto;
     width: 40%;
+    p{
+      font-size: 1.1em;
+      line-height: 1.3;
+    }
+
   }
 }
