+++ src/Components/App/App.tsx
@@ -2,7 +2,7 @@ import ParticleBackground from "../../assets/ParticleBackground";
 import "./App.scss";
 import PopupMenu from "../PopupMenu/PopupMenu";
 import WelcomePage from "../WelcomePage/WelcomePage";
-import AboutMe from "../Perfil/AboutMe";
+import AboutMe from "../AboutMe/AboutMe";
 
 function App() {
   return (
