+++ src/index.scss
@@ -38,6 +38,13 @@ h1 {
   line-height: 1.1;
 }
 
+h2 {
+  font-size: 2em;
+  line-height: 1.1;
+  margin-bottom: 1.5em;
+  margin-top: 0;
+}
+
 html {
   scroll-snap-type: y mandatory;
 }
